+++ NEWS.md
@@ -88,7 +88,7 @@ If running Workbench behind a proxy server, you may need to update your `NO_PROX
 
 ### Dependencies
 - Updated GWT to version 2.10.1 (#15011)
-- Updated Electron to version 31.5.0 (#14982; Desktop)
+- Updated Electron to version 31.7.1 (#14982; Desktop)
 
 ### Deprecated / Removed
 - Removed user preference for turning off focus indicator rectangles (#14352)